reorganize file structure, mock fetch call in app test
