+++ src/Favorites/index.js
