+++ src/Favorites/styles.css
