+++ src/Nav/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import ContentButton from './ContentButton/index.js';
-import Favorites from './Favorites/index.js';
+import ContentButton from '../ContentButton/index.js';
+import Favorites from '../Favorites/index.js';
 import PropTypes from 'prop-types';
 import './styles.css';
