+++ src/index.js
@@ -2,5 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import retrieveScroll from './helpers/retrieveScroll.js';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+
+ReactDOM.render(<App retrieveScroll={retrieveScroll}/>, document.getElementById('root'));
